Refactor Notify Service and TS
